// source/src/shared/lib/enums/ExampleTypes.ts
export enum ExampleTypes {
    INT = "INT",
    FRACTION = "FRACTION",
    MODULE = "MODULE",
    FACTORIAL = "FACTORIAL",
    SQRT = "SQRT",
    DEGREE = "DEGREE",
    SIN = "SIN",
    COS = "COS",
    TAN = "TAN",
    CTG = "CTG",
    ARCSIN = "ARCSIN",
    ARCCOS = "ARCCOS",
    ARCTG = "ARCTG",
    ARCCTG = "ARCCTG",
    LOG = "LOG",
    LG = "LG",
    LN = "LN",
    MATRIX = "MATRIX",
    SIGN = "SIGN",
    EQUALS = "EQUALS",
    PARENTHESIS = "PARENTHESIS",
    CURSOR = "CURSOR"
}